Selva headers cleanup
Require fewer #includes for basic features of Selva.
