+++ server/modules/selva/module/config.c
@@ -1,8 +1,7 @@
 #include <string.h>
 #include <stddef.h>
 #include <redismodule.h>
-#include "cdefs.h"
-#include "errors.h"
+#include "selva.h"
 #include "modinfo.h"
 #include "config.h"
