+++ server/modules/selva/module/edge/edge.c
@@ -4,8 +4,8 @@
 #include "redismodule.h"
 #include "jemalloc.h"
 #include "auto_free.h"
-#include "errors.h"
 #include "svector.h"
+#include "selva.h"
 #include "hierarchy.h"
 #include "selva_object.h"
 #include "subscriptions.h"
