+++ server/modules/selva/module/hierarchy/hierarchy.c
@@ -12,9 +12,7 @@
 #include "selva.h"
 #include "alias.h"
 #include "async_task.h"
-#include "cdefs.h"
 #include "rms.h"
-#include "errors.h"
 #include "edge.h"
 #include "modify.h"
 #include "rpn.h"
