+++ server/modules/selva/module/hierarchy/hierarchy_inactive.c
@@ -1,7 +1,6 @@
 #include <sys/mman.h>
 #include <unistd.h>
 #include "selva.h"
-#include "errors.h"
 #include "hierarchy.h"
 #include "hierarchy_inactive.h"
