+++ server/modules/selva/module/ida.c
@@ -4,7 +4,7 @@
  */
 #include "jemalloc.h"
 #include "bitmap.h"
-#include "errors.h"
+#include "selva.h"
 #include "ida.h"
 
 struct ida {
