+++ server/modules/selva/module/modify.c
@@ -11,20 +11,15 @@
 #include <unistd.h>
 #include "redismodule.h"
 #include "jemalloc.h"
-#include "cdefs.h"
-#include "errors.h"
-#include "async_task.h"
 #include "bitmap.h"
-#include "cdefs.h"
+#include "selva.h"
+#include "async_task.h"
 #include "comparator.h"
 #include "config.h"
 #include "cstrings.h"
-#include "errors.h"
-#include "hierarchy.h"
 #include "libdeflate.h"
-#include "modify.h"
+#include "hierarchy.h"
 #include "rms.h"
-#include "selva.h"
 #include "selva_object.h"
 #include "selva_onload.h"
 #include "selva_set.h"
