+++ server/modules/selva/module/resolve.c
@@ -3,7 +3,6 @@
 #include "selva.h"
 #include "alias.h"
 #include "arg_parser.h"
-#include "errors.h"
 #include "hierarchy.h"
 #include "selva_onload.h"
 #include "subscriptions.h"
