+++ server/modules/selva/module/selva_lang.c
@@ -3,10 +3,9 @@
 #include <string.h>
 #include <strings.h>
 #include "redismodule.h"
+#include "selva.h"
 #include "jemalloc.h"
 #include "selva_onload.h"
-#include "cdefs.h"
-#include "errors.h"
 #include "selva_lang.h"
 #include "selva_object.h"
