+++ server/modules/selva/module/selva_type.c
@@ -1,9 +1,7 @@
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include "cdefs.h"
 #include "redismodule.h"
-#include "errors.h"
 #include "selva.h"
 
 /*
