+++ server/modules/selva/test/hierarchy-mock.c
@@ -1,6 +1,6 @@
 #include <stddef.h>
 #include "hierarchy.h"
-#include "errors.h"
+#include "selva.h"
 
 int SelvaHierarchy_IsNonEmptyField(const struct SelvaHierarchyNode *node, const char *field_str, size_t field_len) {
     return SELVA_ENOENT;
