+++ server/modules/selva/test/hierarchy_detached-mock.c
@@ -1,5 +1,5 @@
 #include "redismodule.h"
-#include "errors.h"
+#include "selva.h"
 #include "hierarchy.h"
 #include "../module/hierarchy/hierarchy_detached.h"
