+++ server/modules/selva/test/units/test-hierarchy.c
@@ -6,7 +6,6 @@
 #include "traversal.h"
 #include "hierarchy.h"
 #include "cdefs.h"
-#include "errors.h"
 #include "../hierarchy-utils.h"
 
 static void setup(void)
