+++ server/modules/selva/test/units/test-selva_set.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include "cdefs.h"
 #include "redismodule.h"
-#include "errors.h"
+#include "selva.h"
 #include "selva_set.h"
 
 static void setup(void)
